+++ R/rmd_to_md.R
@@ -30,14 +30,20 @@
 #'   fig_url_dir = "/software/gcube/",
 #'   title = "2. Occurrence process",
 #'   sidebar_label = "occurrence-process",
-#'   sidebar_order = 2
+#'   sidebar_order = 2,
+#'   replace = c(
+#'     "### Changing number of occurrences over time" =
+#'     "### How to change the number of occurrences over time",
+#'     "man/figures/logo.png" = "https://b-cubed-eu.github.io/gcube/logo.png"
+#'   )
 #' )
 #'
 #' # Clean up (don't do this if you want to keep your files)
 #' unlink("output", recursive = TRUE)
 #' }
 rmd_to_md <- function(rmd_file, md_dir, fig_dir, fig_url_dir, title = NULL,
-                      sidebar_label = NULL, sidebar_order = NULL, logo = NULL) {
+                      sidebar_label = NULL, sidebar_order = NULL,
+                      replace = NULL) {
   # Get the basename of the input rmd_file without extension.
   md_name <- fs::path_file(fs::path_ext_remove(rmd_file))
 
@@ -46,16 +52,15 @@ rmd_to_md <- function(rmd_file, md_dir, fig_dir, fig_url_dir, title = NULL,
     # Store the rmd_file in a subdir of the OS tempdir
     temp_dir <- fs::path_temp("rmd_file")
     # Create the temp_dir if it doesn't exist from an earlier run
-    if (!fs::dir_exists(temp_dir)) {fs::dir_create(temp_dir)}
-    temp_rmd_path <-
-      fs::file_temp(pattern = md_name,
-                    tmp_dir = temp_dir,
-                    ext = ".Rmd")
+    if (!fs::dir_exists(temp_dir)) { fs::dir_create(temp_dir) }
+    temp_rmd_file <- fs::file_temp(
+      pattern = md_name,
+      tmp_dir = temp_dir,
+      ext = ".Rmd"
+    )
 
-    utils::download.file(rmd_file,
-                         temp_rmd_path)
-
-    input_file <- temp_rmd_path
+    utils::download.file(rmd_file, temp_rmd_file)
+    input_file <- temp_rmd_file
 
   } else {
     # The file is local.
@@ -63,9 +68,10 @@ rmd_to_md <- function(rmd_file, md_dir, fig_dir, fig_url_dir, title = NULL,
   }
 
   # Set output
-  md_file_path <- fs::path_ext_set(
+  md_file <- fs::path_ext_set(
     path = fs::path(md_dir, md_name),
-    ext = ".md")
+    ext = ".md"
+  )
 
   # Create directory for markdown file
   if (!fs::dir_exists(md_dir)) {
@@ -93,17 +99,17 @@ rmd_to_md <- function(rmd_file, md_dir, fig_dir, fig_url_dir, title = NULL,
   # Knit
   knitr::knit(
     input = input_file,
-    output = md_file_path
+    output = md_file
   )
 
   # Update front matter
   update_frontmatter(
-    md_file_path,
+    md_file,
     rmd_file,
     title = title,
     sidebar_label = sidebar_label,
     sidebar_order = sidebar_order,
-    logo = logo
+    replace = replace
   )
 
   # Reset knitting options to the original settings